# refine barret multiplier