--- a/icicle/CMakeLists.txt
+++ b/icicle/CMakeLists.txt
@@ -20,7 +20,6 @@ else()
 endif()
 
 
-
 include(cmake/field.cmake)
 include(cmake/curve.cmake)
 include(cmake/hash.cmake)
@@ -29,9 +28,6 @@ include(cmake/hash.cmake)
 if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE "Release" CACHE STRING "Choose the type of build: Debug, Release, RelWithDebInfo, MinSizeRel." FORCE)
 endif()
-
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Og -g")
-set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -Og -g")
 
 # Print the selected build type
 message(STATUS "Build type: ${CMAKE_BUILD_TYPE}")