--- a/icicle/CMakeLists.txt
+++ b/icicle/CMakeLists.txt
@@ -20,6 +20,7 @@ else()
 endif()
 
 
+
 include(cmake/field.cmake)
 include(cmake/curve.cmake)
 include(cmake/hash.cmake)