This PR improves the barret multiplier by ~25% with a few tweaks:

- some more inlining
- loop ordering

As a result **_(Measured for BN254 single core on M1-pro)_**:

- scalar multiplication (1M elements):  ~27ms -> ~21.5ms (~1.25x) (Montgomery PR is  ~16.5ms **~1.3x faster**)
- EC adder (1K elements): ~0.36ms -> ~0.325ms (~1.1x) (Montgomery PR is  ~0.36ms **~1.1x slower**)
- MSM logn=12 (single core): 53ms --> 48ms (~1.1x) (Montgomery PR is ~53ms **~1.1x slower**)
- NTT logn=16 (single core): ~187ms --> ~165ms (~1.13x) (Montgomery PR is ~160ms **~1.03x faster**)
- ECNTT logn=10 (single core): ~ 402ms--> ~350ms (~1.15x) (Montgomery PR is ~ 350ms **~1x**)